+++ src/main/java/uk/gov/hmcts/reform/hmc/controllers/LinkedHearingGroupController.java
@@ -7,6 +7,7 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import uk.gov.hmcts.reform.hmc.service.AccessControlService;
 import uk.gov.hmcts.reform.hmc.service.LinkedHearingGroupService;
 
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
@@ -15,9 +16,12 @@ import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 public class LinkedHearingGroupController {
 
     private LinkedHearingGroupService linkedHearingGroupService;
+    private AccessControlService accessControlService;
 
-    public LinkedHearingGroupController(LinkedHearingGroupService linkedHearingGroupService) {
+    public LinkedHearingGroupController(LinkedHearingGroupService linkedHearingGroupService,
+                                        AccessControlService accessControlService) {
         this.linkedHearingGroupService = linkedHearingGroupService;
+        this.accessControlService = accessControlService;
     }
 
     @DeleteMapping(path = "/linkedHearingGroup/{id}", consumes = APPLICATION_JSON_VALUE,
@@ -30,6 +34,7 @@ public class LinkedHearingGroupController {
         @ApiResponse(code = 500, message = "Error occurred on the server")
     })
     public void deleteHearingGroup(@PathVariable("id") Long hearingGroupId) {
+        accessControlService.verifyCaseAccess(null);
         linkedHearingGroupService.deleteLinkedHearingGroup(hearingGroupId);
     }
 }
